Removed NotificationList
